requirements.txt
pytest
allure-pytest
selene == 2.0.0rc9
Appium-Python-Client==4.0.0
python-dotenv
requests
pydantic
pydantic-settings==2.10.1